+++ packages/babel-preset-expo/src/common.ts
@@ -53,7 +53,8 @@ export function getIsDev(caller: any) {
 }
 
 export function getIsFastRefreshEnabled(caller: any) {
-  return caller?.isFastRefreshEnabled ?? false;
+  if (!caller) return false;
+  return caller.isHMREnabled && !caller.isServer && !caller.isNodeModule && getIsDev(caller);
 }
 
 export function getIsProd(caller: any) {
@@ -61,6 +62,10 @@ export function getIsProd(caller: any) {
 
   // https://babeljs.io/docs/options#envname
   return process.env.BABEL_ENV === 'production' || process.env.NODE_ENV === 'production';
+}
+
+export function getIsNodeModule(caller: any): boolean {
+  return caller?.isNodeModule ?? false;
 }
 
 export function getBaseUrl(caller: any): string {
@@ -75,8 +80,9 @@ export function getInlineEnvVarsEnabled(caller: any): boolean {
   const isWebpack = getBundler(caller) === 'webpack';
   const isDev = getIsDev(caller);
   const isServer = getIsServer(caller);
+  const isNodeModule = getIsNodeModule(caller);
   const preserveEnvVars = caller?.preserveEnvVars;
   // Development env vars are added in the serializer to avoid caching issues in development.
   // Servers have env vars left as-is to read from the environment.
-  return !isWebpack && !isDev && !isServer && !preserveEnvVars;
+  return !isNodeModule && !isWebpack && !isDev && !isServer && !preserveEnvVars;
 }