+++ packages/babel-preset-expo/src/index.ts
@@ -63,6 +63,7 @@ function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): 
   const isDev = api.caller(getIsDev);
   const isFastRefreshEnabled = api.caller(getIsFastRefreshEnabled);
   const baseUrl = api.caller(getBaseUrl);
+
   // Unlike `isDev`, this will be `true` when the bundler is explicitly set to `production`,
   // i.e. `false` when testing, development, or used with a bundler that doesn't specify the correct inputs.
   const isProduction = api.caller(getIsProd);
@@ -172,7 +173,13 @@ function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): 
   }
 
   if (isFastRefreshEnabled) {
-    extraPlugins.push(require('react-refresh/babel'));
+    extraPlugins.push([
+      require('react-refresh/babel'),
+      {
+        // We perform the env check to enable `isFastRefreshEnabled`.
+        skipEnvCheck: true,
+      },
+    ]);
   }
 
   return {